cgan/requirements.txt
torch==2.8.0
torchvision==0.23.0
pandas==2.3.2
openpyxl==3.1.5
pillow==11.3.0
tqdm==4.67.1
matplotlib==3.10.6
scikit-image==0.25.2
torchmetrics==1.8.2
lpips==0.1.4
torchmetrics[image]
